Added weapon skill to miss calculation.
Added calculation for weapon skill to miss chance based on https://forum.elysium-project.org/topic/47501-spreadsheet-warrior-naxxaq-bis/

It will use current target level with the assumption that they have level*5 defense.
If no target is found, or it is a boss, it assumes (your level + 3) * 5 for defense of the target (a boss)

Still unsure of some of the values in here, I made this 4 years ago one night, and for some reason people still use it so I got asked to update this. But, unsure if all the values here are still correct.